+++ tests/end2end/utils/utils.py
@@ -3,6 +3,7 @@
 # pylint: disable=import-error
 # pylint: disable=no-name-in-module
 import logging
+import os
 from enum import Enum
 from time import sleep
 
@@ -16,7 +17,6 @@ from qiboconnection.typings.connection import ConnectionConfiguration
 from qiboconnection.typings.enums import DeviceAvailability, DeviceStatus, JobStatus
 from qiboconnection.typings.job_data import JobData
 
-from .http import get_api, get_logging_conf
 from .operations import is_development
 
 
@@ -558,3 +558,53 @@ def admin_release_device(device: Device, api: API):
         pytest.skip(f"{device.name} is {device._status}")
 
     api.release_device(device_id=device.id)
+
+
+def get_logging_conf(role: UserRole = UserRole.ADMIN) -> ConnectionConfiguration:
+    """Build a ConnectionConfiguration object from the keys defined in environment
+
+    Args:
+        role (UserRoles): admin, bsc_user or qilimanjaro_user
+
+    Raises:
+        MissingCredentialsException: missing or wrong credentials
+
+    Returns:
+        ConnectionConfiguration: instance with credentials for creating API instance
+    """
+
+    if role == UserRole.ADMIN:
+        public_login_username = os.getenv("PUBLIC_LOGIN_ADMIN_USERNAME")
+        public_login_key = os.getenv("PUBLIC_LOGIN_ADMIN_KEY")
+
+    elif role == UserRole.BSC:
+        public_login_username = os.getenv("PUBLIC_LOGIN_BSC_USERNAME")
+        public_login_key = os.getenv("PUBLIC_LOGIN_BSC_KEY")
+
+    elif role == UserRole.QILI:
+        public_login_username = os.getenv("PUBLIC_LOGIN_QILI_USERNAME")
+        public_login_key = os.getenv("PUBLIC_LOGIN_QILI_KEY")
+
+    elif role == UserRole.MACHINE:
+        public_login_username = os.getenv("PUBLIC_LOGIN_MACHINE_USERNAME")
+        public_login_key = os.getenv("PUBLIC_LOGIN_MACHINE_KEY")
+
+    # previous try/except was not catching the error
+    if public_login_username is None or public_login_key is None:
+        raise MissingCredentialsException(
+            "Login failed. Credentials were not provided in the environment or are not valid. Recall you need to define a environment variables for every user role."
+        )
+
+    return ConnectionConfiguration(username=public_login_username, api_key=public_login_key)
+
+
+def get_api(logging_conf: ConnectionConfiguration) -> API:
+    """Returns API instance given a ConnectionConfiguration instance
+
+    Args:
+        logging_conf: login credentials
+
+    Returns:
+        API: api usable api instance
+    """
+    return API(configuration=logging_conf)
